requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
boltons==19.3.0
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.4
Django==2.2.5
django-cors-headers==3.1.0
django-js-asset==1.2.2
django-mptt==0.10.0
django-rest-auth==0.9.5
djangorestframework==3.10.3
drf-writable-nested==0.5.3
drf-yasg==1.17.0
entrypoints==0.3
flake8==3.7.8
hypothesis==4.36.2
hypothesis-jsonschema==0.9.8
idna==2.8
importlib-metadata==0.23
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
jsonschema==3.0.2
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.0
psycopg2==2.8.4
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.2
pyrsistent==0.15.4
pytest==5.1.3
pytz==2019.2
PyYAML==5.1.2
redis==3.2.1
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.1.2
schemathesis==0.5.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
wcwidth==0.1.7
zipp==0.6.0
sphinxapi-py3==2.1.11
SolrClient==0.3.1
Whoosh==2.7.4